viewer logfile
